requirements.txt
taichi==1.2.2
numpy==1.23.1
matplotlib
tqdm
pandas
hydra-core